requirements.txt
# Core Django stack
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0

# Async tasks
celery==5.4.0
redis==5.2.1

# Env / config
python-dotenv==1.2.1
dotenv==0.9.9

# AI / ML / embeddings
numpy==2.3.5
sentence-transformers==5.1.2
torch==2.9.1
transformers==4.57.3
scikit-learn==1.7.2
scipy==1.16.3
Pillow==12.0.0

# LLM client (Groq)
groq==0.37.1

# Payments
razorpay==2.0.0

# HTTP & helpers (explicitly pinned, though they’re deps)
requests==2.32.5
httpx==0.28.1
anyio==4.12.0
typing-extensions==4.15.0
